Move autofillAddressPanel.js and autofillCreditCardPanel.js from js to app folder
Auditors:

Test Plan: n/a
