+++ js/components/main.js
@@ -30,8 +30,8 @@ const BraveryPanel = require('./braveryPanel')
 const ClearBrowsingDataPanel = require('./clearBrowsingDataPanel')
 const ImportBrowserDataPanel = require('../../app/renderer/components/importBrowserDataPanel')
 const WidevinePanel = require('../../app/renderer/components/widevinePanel')
-const AutofillAddressPanel = require('./autofillAddressPanel')
-const AutofillCreditCardPanel = require('./autofillCreditCardPanel')
+const AutofillAddressPanel = require('../../app/renderer/components/autofill/autofillAddressPanel')
+const AutofillCreditCardPanel = require('../../app/renderer/components/autofill/autofillCreditCardPanel')
 const AddEditBookmark = require('../../app/renderer/components/bookmarks/addEditBookmark')
 const LoginRequired = require('../../app/renderer/components/loginRequired')
 const ReleaseNotes = require('../../app/renderer/components/releaseNotes')
